Raw XRD generation: Merge same type input objects
Generating "raw" XRDs is based on extracting non-empty values of each
terraform object. In case when there are multiple objects of the same type,
the set of object parameters needs to be merged (combined) into a single
object, otherwise the resulting XRD would be missing some fields that
were filled only in some of the input objects.