Fix merging of arrays within spec values
This fixes an issue with obtaining a generalized XRD based on a set of
input resources, where some of subsequent resources contain an empty
array-of-objects value, effectively "forgetting" the structure inside
such array.

Here a simple example of a set of objects that would exhibit such issue:

```
[
  { "v": 10, "a": [{ "xxx": 0}] },
  { "v": 2 },
  { "v": 5 },
  { "v": 3, "a": [] }
]
```

In previous implementation, the object schema identification would
operate on the following merged object:
```
{ "v": 3, "a": [] }
```

Which led to ignoring the structure of the elements of `a`.

The fixed implementation produces a more properly merged object:
```
{ "v": 3, "a": [ {"xxx":0} ] }
```

Note that this problem was only exhibited if `a` was an empty array on
one of the objects further in the input, effectively overriding the
possible structure detection done for an earlier object. This means
there could be an alternative solution to the problem: one in which
a schema is first identified for all objects and then the schemas
(instead of input objects) get merged. That however seemed more
complicated and less efficient.